all database functions moved to m_database.h
git-svn-id: http://svn.miranda-ng.org/main/trunk@12385 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/SimpleStatusMsg/src/options.cpp
@@ -942,7 +942,7 @@ INT_PTR CALLBACK DlgOptionsProc(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM l
 
 						if (data->proto_msg[j].flags & PROTO_THIS_MSG)
 						{
-							int len = mir_tstrlen(data->proto_msg[j].msg);
+							size_t len = mir_tstrlen(data->proto_msg[j].msg);
 							if (len > 0)
 							{	
 								if (data->proto_msg[k+1].msg == NULL)
